+++ src/main/java/com/kamijoucen/ruler/value/convert/IntegerConvert.java
@@ -12,12 +12,12 @@ public class IntegerConvert implements ValueConvert {
 
     @Override
     public BaseValue realToBase(Object value) {
-        return new IntegerValue(Integer.parseInt(value.toString()));
+        return new IntegerValue(Long.parseLong(value.toString()));
     }
 
     @Override
     public Object baseToReal(BaseValue value) {
-        return Integer.parseInt(value.toString());
+        return Long.parseLong(value.toString());
     }
 
 }
